airflow/requirements.txt
# Core Airflow
apache-airflow==3.0.2

# Core Provider Packages
apache-airflow-providers-common-compat==1.12.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.30.3
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.10.3

# Compression
py7zr==1.0.0

# HTTP Requests
requests>=2.31.0

# Cloud Providers
apache-airflow-providers-amazon>=8.0.0
boto3>=1.26.0

# Database Providers
apache-airflow-providers-databricks==7.6.0
apache-airflow-providers-snowflake==5.7.1
apache-airflow-providers-mysql==6.4.2

# Integration Providers
apache-airflow-providers-airbyte==5.3.2


# ✅ ADD THESE - Soda packages
soda-core-spark==3.3.2
soda-core-postgres==3.3.2
soda-core-mysql==3.3.2